Split up Travertino style tests
